requirements.txt
Django==4.0.4
Jinja2==3.1.1
MarkupSafe==2.1.1
PyJWT==2.3.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-static==0.0.6
django-environ==0.8.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
itypes==1.2.0
oauthlib==3.2.0
pip	21.1.2==22.0.4
pycparser==2.21
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
setuptools==57.0.0
six	1.16.0==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wheel==0.36.2